pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "any2any"
version = "0.1.0"
description = "Convert arbitrary files from one format to another"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "Pillow>=10.0",
    "pillow-heif>=0.16.0",
    "rawpy>=0.19.0",
    "numpy>=1.24.0",
    "psd-tools>=1.9.0",
]

[project.optional-dependencies]
svg = [
    "svglib>=0.19.0",
    "reportlab>=4.0",
]
jxl = [
    "pillow-jxl-plugin>=1.0.0",
]
all = [
    "any2any[svg,jxl]",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
any2any = "any2any.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
